Added processing of CP and SEMESTER restrictions. Added also messaging and fixed persistence and JMX in messaging configs. Extended Services, Senders and Repos. Extended attributes of Module and PORestriction.
... now I'm officially tired.

+++ src/main/java/de/hsrm/mi/swtp/exchangeplatform/model/data/PORestriction.java
@@ -74,7 +74,7 @@ public class PORestriction implements Model {
 		@Schema(name = "min_semesters", defaultValue = "0", nullable = true, description = "All modules have to be passed from the first semester up until this given (inclusive).")
 		@Column(nullable = false, name = "min_semesters")
 		@JsonProperty(value = "min_semesters", defaultValue = "0")
-		Integer minSemesters = 0;
+		Long minSemesters = 0L;
 	}
 
 	@Entity
